Created SouffleType header file, containing Kind.
