+++ src/AnalysisType.h
@@ -17,6 +17,7 @@
 #pragma once
 
 #include "AstType.h"
+#include "SouffleType.h"
 #include "Util.h"
 #include <algorithm>
 #include <cassert>
@@ -29,25 +30,6 @@
 #include <vector>
 
 namespace souffle {
-
-// TODO: move to new file
-enum class Kind { SYMBOL, NUMBER, RECORD };
-
-inline std::ostream& operator<<(std::ostream& os, Kind kind) {
-    switch (kind) {
-        case Kind::SYMBOL:
-            os << "symbol";
-            break;
-        case Kind::NUMBER:
-            os << "number";
-            break;
-        case Kind::RECORD:
-            os << "any record";
-            break;
-    }
-
-    return os;
-}
 
 /** Represents a node in the type lattice during type analysis */
 class AnalysisType {
